pyproject.toml
[tool.poetry]
name = "vectordb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sentence-transformers = "^3.0.1"
qdrant-client = "^1.11.1"
pybind11 = "^2.13.5"
numpy = "<2.0"
transformers = "^4.44.2"
torch = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
